Proper minor mode hooking, re-worked implementations
The existing package was activating a pre-command hook globally and then inside
that hook, inspecting the state of the minor mode to determine if it should log
or not.  Since minor modes can activate hooks inside their bodies, this was
unecessary and also had the side effect of being pretty dirty.

By going the proper way and then factoring out some common routines, the results
are getting cleaner.

There was some mixing of the duties of routines and commands as well, and so it
was difficult to decide when to honor the settings to automatically show the
buffer or stop logging etc.

Storing the buffer in a variable was eliminated since it's too easy to
accidentally kill that buffer and then have hooks pre-command get into an
inconsistent state later as they fail and are removed.